pipeline/requirements.txt
transformers==4.40.0
torch==2.2.2
bertopic==0.16.2
sentence-transformers==2.7.0
xgboost==2.0.3
lightgbm==4.3.0
scikit-learn==1.4.2
mlflow==2.11.1
duckdb==0.10.1
pandas==2.2.2
numpy==1.26.4
pyarrow==16.0.0
shap==0.45.0
vectorbt==0.26.1
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
click==8.1.7
loguru==0.7.2
tqdm==4.66.2
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0
scipy==1.13.0
statsmodels==0.14.2
